lib/python/pyproject.toml
[tool.poetry]
name = "watermark"
version = "0.1.0"
description = ""
authors = ["Alice Xie <alicez.xie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.1"
pywavelets = "^1.4.1"
pillow = "^9.4.0"
opencv-python = "^4.7.0.68"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
